# Lab9/backend/src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Your API Title",
        "description": "Your API Description",
        "contact": {
            "email": "your.email@example.com"
        },
        "license": {
            "name": "MIT License",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/programs": {
            "get": {
                "tags": [
                    "Programs"
                ],
                "summary": "Get list of programs",
                "description": "Returns a list of programs",
                "operationId": "c3bb77153f2117d65a80a27eb16cbce1",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Program"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Programs"
                ],
                "summary": "Create a new program",
                "operationId": "storeProgram",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Name of the program",
                                        "type": "string"
                                    },
                                    "duration": {
                                        "description": "Duration of the program",
                                        "type": "integer"
                                    },
                                    "version": {
                                        "description": "Version of the program",
                                        "type": "string",
                                        "maxLength": 4
                                    },
                                    "major_id": {
                                        "description": "ID of the major",
                                        "type": "integer"
                                    },
                                    "program_type_id": {
                                        "description": "Type ID of the program",
                                        "type": "integer"
                                    },
                                    "valid_from": {
                                        "description": "Valid from date",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Program created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Program"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Invalid input"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/programs/{program}": {
            "get": {
                "tags": [
                    "Programs"
                ],
                "summary": "Get a program by ID",
                "description": "Returns a single program",
                "operationId": "getProgramById",
                "parameters": [
                    {
                        "name": "program",
                        "in": "path",
                        "description": "ID of the program to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Program"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Program not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Program not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Programs"
                ],
                "summary": "Update a program by ID",
                "description": "Updates the details of a specific program",
                "operationId": "updateProgramById",
                "parameters": [
                    {
                        "name": "program",
                        "in": "path",
                        "description": "ID of the program to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Name of the program",
                                        "type": "string"
                                    },
                                    "duration": {
                                        "description": "Duration of the program",
                                        "type": "integer"
                                    },
                                    "version": {
                                        "description": "Version of the program",
                                        "type": "string",
                                        "maxLength": 4
                                    },
                                    "major_id": {
                                        "description": "ID of the major",
                                        "type": "integer"
                                    },
                                    "program_type_id": {
                                        "description": "Type ID of the program",
                                        "type": "integer"
                                    },
                                    "valid_from": {
                                        "description": "Valid from date",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Program"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Invalid input"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Program not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Program not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Programs"
                ],
                "summary": "Delete a program by ID",
                "description": "Deletes a specific program",
                "operationId": "deleteProgramById",
                "parameters": [
                    {
                        "name": "program",
                        "in": "path",
                        "description": "ID of the program to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Program deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Program not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Program not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/programs/{program}/edit": {
            "get": {
                "tags": [
                    "Programs"
                ],
                "summary": "Edit a program by ID",
                "description": "Returns the data of a single program for editing",
                "operationId": "editProgramById",
                "parameters": [
                    {
                        "name": "program",
                        "in": "path",
                        "description": "ID of the program to edit",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Program"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Program not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Program not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "Program": {
                "title": "Program",
                "description": "Program model schema",
                "required": [
                    "name",
                    "program_type_id"
                ],
                "properties": {
                    "name": {
                        "description": "Name of the program",
                        "type": "string",
                        "example": "Computer Science"
                    },
                    "duration": {
                        "description": "Duration of the program in years",
                        "type": "integer",
                        "example": 4
                    },
                    "version": {
                        "description": "Version of the program",
                        "type": "string",
                        "example": "1.0"
                    },
                    "major_id": {
                        "description": "ID of the major associated with the program",
                        "type": "integer",
                        "example": 1
                    },
                    "program_type_id": {
                        "description": "ID of the program type",
                        "type": "integer",
                        "example": 2
                    },
                    "valid_from": {
                        "description": "Date from which the program is valid",
                        "type": "string",
                        "format": "date",
                        "example": "2022-01-01"
                    }
                },
                "type": "object"
            }
        }
    }
}